+++ fixed-vector/ChangeLog.md
@@ -25,6 +25,10 @@ Changes in 1.3.0.0
 
   * `Data.Vector.Fixed.sequenceA` is deprecated in favor of `sequence`
 
+  * `foldl'` and `ifoldl'` functions added.
+
+  * Implement `sum` as in terms of `foldl'`.
+
 
 Changes in 1.2.3.0